Show all expressions as polynomials
