+++ acvm-repo/acir/src/native_types/expression/mod.rs
@@ -51,23 +51,7 @@ impl<F: AcirField> Default for Expression<F> {
 
 impl<F: AcirField> std::fmt::Display for Expression<F> {
     fn fmt(&self, f: &mut std::fmt::Formatter) -> std::fmt::Result {
-        if let Some(witness) = self.to_witness() {
-            return write!(f, "{witness}");
-        }
-
-        write!(f, "EXPR [ ")?;
-        for i in &self.mul_terms {
-            write!(f, "({}, {}, {}) ", i.0, i.1, i.2)?;
-        }
-        for i in &self.linear_combinations {
-            write!(f, "({}, {}) ", i.0, i.1)?;
-        }
-
-        let q_c_as_string = self.q_c.to_string();
-        if q_c_as_string != "0" {
-            write!(f, "{} ", self.q_c)?;
-        }
-        write!(f, "]")
+        display_expression(self, false, f)
     }
 }
 
@@ -382,6 +366,132 @@ impl<F: AcirField> From<Witness> for Expression<F> {
             mul_terms: Vec::new(),
         }
     }
+}
+
+/// Displays an expression as a quadratic polynomial.
+/// If `as_equal_to_zero` is true, the expression is displayed as equaling zero,
+/// where it's tried to shown as a polynomial equal to the largest witness, if possible.
+pub(crate) fn display_expression<F: AcirField>(
+    expr: &Expression<F>,
+    as_equal_to_zero: bool,
+    f: &mut std::fmt::Formatter<'_>,
+) -> std::fmt::Result {
+    // This is set to an index if we show this expression "as a witness assignment", meaning
+    // that the linear combination at this index must not be printed again.
+    let mut assignment_witness: Option<usize> = None;
+
+    // If true, negate all coefficients when printing.
+    // This is set to true if we show this expression "as a witness assignment", and the witness
+    // had a coefficient of 1 and we "moved" everything to the right of the equal sign.
+    let mut negate_coefficients = false;
+
+    // Find a linear combination with a coefficient of 1 or -1 and, if there are many,
+    // keep the one with the largest witness.
+    let linear_witness_one = if as_equal_to_zero {
+        expr.linear_combinations
+            .iter()
+            .enumerate()
+            .filter(|(_, (coefficient, _))| coefficient.is_one() || (-*coefficient).is_one())
+            .max_by_key(|(_, (_, witness))| witness)
+    } else {
+        None
+    };
+
+    // If we find one, show the expression as equaling this witness to everything else
+    // (this is likely to happen as in ACIR gen we tend to equate a witness to previous expressions)
+    if let Some((index, (coefficient, witness))) = linear_witness_one {
+        assignment_witness = Some(index);
+        negate_coefficients = coefficient.is_one();
+        write!(f, "{witness} = ")?;
+    } else if as_equal_to_zero {
+        write!(f, "0 = ")?;
+    }
+
+    let mut printed_term = false;
+
+    for (coefficient, witness1, witness2) in &expr.mul_terms {
+        let witnesses = [*witness1, *witness2];
+        display_term(*coefficient, witnesses, printed_term, negate_coefficients, f)?;
+        printed_term = true;
+    }
+
+    for (index, (coefficient, witness)) in expr.linear_combinations.iter().enumerate() {
+        if assignment_witness
+            .is_some_and(|show_as_assignment_index| show_as_assignment_index == index)
+        {
+            // We already printed this term as part of the assignment
+            continue;
+        }
+
+        let witnesses = [*witness];
+        display_term(*coefficient, witnesses, printed_term, negate_coefficients, f)?;
+        printed_term = true;
+    }
+
+    if expr.q_c.is_zero() {
+        if as_equal_to_zero && !printed_term {
+            write!(f, "0")?;
+        }
+    } else {
+        let coefficient = expr.q_c;
+        let coefficient = if negate_coefficients { -coefficient } else { coefficient };
+        let coefficient_as_string = coefficient.to_string();
+        let coefficient_is_negative = coefficient_as_string.starts_with('-');
+
+        if printed_term {
+            if coefficient_is_negative {
+                write!(f, " - ")?;
+            } else {
+                write!(f, " + ")?;
+            }
+        }
+
+        let coefficient =
+            if printed_term && coefficient_is_negative { -coefficient } else { coefficient };
+        write!(f, "{coefficient}")?;
+    }
+
+    Ok(())
+}
+
+fn display_term<F: AcirField, const N: usize>(
+    coefficient: F,
+    witnesses: [Witness; N],
+    printed_term: bool,
+    negate_coefficients: bool,
+    f: &mut std::fmt::Formatter<'_>,
+) -> std::fmt::Result {
+    let coefficient = if negate_coefficients { -coefficient } else { coefficient };
+    let coefficient_as_string = coefficient.to_string();
+    let coefficient_is_negative = coefficient_as_string.starts_with('-');
+
+    if printed_term {
+        if coefficient_is_negative {
+            write!(f, " - ")?;
+        } else {
+            write!(f, " + ")?;
+        }
+    }
+
+    let coefficient =
+        if printed_term && coefficient_is_negative { -coefficient } else { coefficient };
+
+    if coefficient.is_one() {
+        // Don't print the coefficient
+    } else if (-coefficient).is_one() {
+        write!(f, "-")?;
+    } else {
+        write!(f, "{coefficient}*")?;
+    }
+
+    for (index, witness) in witnesses.iter().enumerate() {
+        if index != 0 {
+            write!(f, "*")?;
+        }
+        write!(f, "{witness}")?;
+    }
+
+    Ok(())
 }
 
 #[cfg(test)]
