+++ acvm-repo/acir/src/native_types/mod.rs
@@ -6,6 +6,7 @@ mod witness_map;
 mod witness_stack;
 
 pub use expression::Expression;
+pub(crate) use expression::display_expression;
 pub use witness::Witness;
 pub use witness_map::WitnessMap;
 pub use witness_stack::WitnessStack;
