+++ compiler/noirc_evaluator/src/acir/tests/call.rs
@@ -60,9 +60,9 @@ fn basic_call_with_outputs_assert(inline_type: InlineType) {
     private parameters: [w0, w1]
     public parameters: []
     return values: []
-    CALL func 1: PREDICATE: EXPR [ 1 ]
+    CALL func 1: PREDICATE: 1
     inputs: [w0, w1], outputs: [w2]
-    CALL func 1: PREDICATE: EXPR [ 1 ]
+    CALL func 1: PREDICATE: 1
     inputs: [w0, w1], outputs: [w3]
     EXPR w3 = w2
 
@@ -103,9 +103,9 @@ fn call_output_as_next_call_input(inline_type: InlineType) {
     private parameters: [w0, w1]
     public parameters: []
     return values: []
-    CALL func 1: PREDICATE: EXPR [ 1 ]
+    CALL func 1: PREDICATE: 1
     inputs: [w0, w1], outputs: [w2]
-    CALL func 1: PREDICATE: EXPR [ 1 ]
+    CALL func 1: PREDICATE: 1
     inputs: [w2, w1], outputs: [w3]
     EXPR w3 = w2
 
@@ -150,9 +150,9 @@ fn basic_nested_call(inline_type: InlineType) {
     private parameters: [w0, w1]
     public parameters: []
     return values: []
-    CALL func 1: PREDICATE: EXPR [ 1 ]
+    CALL func 1: PREDICATE: 1
     inputs: [w0, w1], outputs: [w2]
-    CALL func 1: PREDICATE: EXPR [ 1 ]
+    CALL func 1: PREDICATE: 1
     inputs: [w0, w1], outputs: [w3]
     EXPR w3 = w2
 
@@ -162,7 +162,7 @@ fn basic_nested_call(inline_type: InlineType) {
     public parameters: []
     return values: [w2]
     EXPR w3 = w0 + 2
-    CALL func 2: PREDICATE: EXPR [ 1 ]
+    CALL func 2: PREDICATE: 1
     inputs: [w3, w1], outputs: [w4]
     EXPR w4 = w2
